created main.cpp
Ben and I tried to plan out what to include in the main script that controlled out application. There are notes toward the top of the script that note a few ides.

Ben and I also made some separate notes in other places about how we thought we should structure the overall project that we can consolidate somewhere later today.